ALB-10 Move implementations into packages
The source code was organized more in a library typical way making only
necessary methods and classes visible to the client using the library.

Note that the code was only moved, no changes were performed inside of the
source files, except for the GadgetmanagerFactory class which was newly
implemented to allow the creation of a Gadgetmanager.

Change-Id: I71c5506cc7fd04fa106f1e89a9ed6c2314f55b56